+++ runtime/db/migrations/000006_add_started_at_index.up.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+CREATE INDEX idx_started_at_desc ON inferences (started_at DESC);
+
+COMMIT;